fix: selecionar size obrigatorio
